split out MixerMessageParser into its own class
