+++ ChatCore/Models/Mixer/MixerSentMessageInfo.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Reflection;
+using System.Text;
+
+namespace ChatCore.Models.Mixer
+{
+    public class MixerSentMessageInfo
+    {
+        public Assembly Assembly;
+        public string MessageType;
+    }
+}
