+++ ChatCore/Services/Mixer/MixerService.cs
@@ -14,57 +14,6 @@ using System.Collections.ObjectModel;
 
 namespace ChatCore.Services.Mixer
 {
-    public class MixerMessageParser : IChatMessageParser
-    {
-        public bool ParseRawMessage(string rawMessage, ConcurrentDictionary<string, IChatChannel> channelInfo, IChatUser loggedInUser, out IChatMessage[] parsedMessage)
-        {
-            var parsedMessages = new List<IChatMessage>();
-            parsedMessage = parsedMessages.ToArray();
-            if (string.IsNullOrEmpty(rawMessage))
-            {
-                return false;
-            }
-            var json = JSON.Parse(rawMessage);
-            if(json == null)
-            {
-                return false;
-            }
-            if(!json.TryGetKey("type", out var t))
-            {
-                return false;
-            }
-            var type = t.Value;
-            var messageMeta = new Dictionary<string, string>();
-            foreach(var key in json.Keys)
-            {
-                messageMeta.Add(key, json[key].Value);
-            }
-
-            var mixerMessage = new MixerMessage()
-            {
-                Type = type,
-                Metadata = new ReadOnlyDictionary<string, string>(messageMeta)
-            };
-            switch(type)
-            {
-                case "reply":
-                    mixerMessage.Message = rawMessage;
-                    mixerMessage.Id = json.TryGetKey("id", out var id) ? id.AsInt.ToString() : "";
-                    parsedMessages.Add(mixerMessage);
-                    return true;
-                default:
-                    break;
-            }
-            return false;
-        }
-    }
-
-    public class SentMessageInfo
-    {
-        public Assembly Assembly;
-        public string MessageType;
-    }
-
     public class MixerService : ChatServiceBase, IChatService
     {
         private ConcurrentDictionary<string, IChatChannel> _channels = new ConcurrentDictionary<string, IChatChannel>();
@@ -101,7 +50,7 @@ namespace ChatCore.Services.Mixer
         private bool _isStarted = false;
         private object _messageReceivedLock = new object(), _initLock = new object();
         private CancellationTokenSource _processMessageQueueCancellation = null;
-        private ConcurrentDictionary<string, SentMessageInfo> _sentMessageInfo = new ConcurrentDictionary<string, SentMessageInfo>();
+        private ConcurrentDictionary<string, MixerSentMessageInfo> _sentMessageInfo = new ConcurrentDictionary<string, MixerSentMessageInfo>();
 
         internal async void Start(bool forceReconnect = false)
         {
@@ -269,7 +218,7 @@ namespace ChatCore.Services.Mixer
                     mixerChannel.Socket.SendMessage($"{{\"type\": \"method\", \"method\": \"{method}\", \"arguments\": {rawArguments}, \"id\": {msgId}}}");
                     if (forwardToSharedClients)
                     {
-                        _sentMessageInfo.TryAdd(msgId.ToString(), new SentMessageInfo()
+                        _sentMessageInfo.TryAdd(msgId.ToString(), new MixerSentMessageInfo()
                         {
                             Assembly = assembly,
                             MessageType = method
